Remove useless Delegate base class
