+++ libqtile/layout/base.py
@@ -216,132 +216,6 @@ class Layout(CommandObject, configurable.Configurable, metaclass=ABCMeta):
         pass
 
 
-class Delegate(Layout):
-    """Base for all delegation layouts"""
-
-    def __init__(self, **config):
-        self.layouts = {}
-        Layout.__init__(self, **config)
-
-    def clone(self, group):
-        c = Layout.clone(self, group)
-        c.layouts = {}
-        return c
-
-    @abstractmethod
-    def _get_layouts(self):
-        """Returns all children layouts"""
-        pass
-
-    @abstractmethod
-    def _get_active_layout(self):
-        """Returns layout to which delegate commands to"""
-        pass
-
-    def delegate_layout(self, windows, mapping):
-        """Delegates layouting actual windows
-
-        Parameters
-        ===========
-        windows:
-            windows to layout
-        mapping:
-            mapping from layout to ScreenRect for each layout
-        """
-        grouped = {}
-        for w in windows:
-            lay = self.layouts[w]
-            if lay in grouped:
-                grouped[lay].append(w)
-            else:
-                grouped[lay] = [w]
-        for lay, wins in grouped.items():
-            lay.layout(wins, mapping[lay])
-
-    def remove(self, win):
-        lay = self.layouts.pop(win)
-        focus = lay.remove(win)
-        if not focus:
-            layouts = self._get_layouts()
-            idx = layouts.index(lay)
-            while idx < len(layouts) - 1 and not focus:
-                idx += 1
-                focus = layouts[idx].focus_first()
-        return focus
-
-    @abstractmethod
-    def show(self, screen_rect):
-        """Tells the layout to show itself on the given ScreenRect"""
-        pass
-
-    def hide(self):
-        for lay in self._get_layouts():
-            lay.hide()
-
-    def focus(self, win):
-        self.layouts[win].focus(win)
-
-    def blur(self):
-        for lay in self._get_layouts():
-            lay.blur()
-
-    def focus_first(self):
-        layouts = self._get_layouts()
-        for lay in layouts:
-            win = lay.focus_first()
-            if win:
-                return win
-
-    def focus_last(self):
-        layouts = self._get_layouts()
-        for lay in reversed(layouts):
-            win = lay.focus_last()
-            if win:
-                return win
-
-    def focus_next(self, win):
-        layouts = self._get_layouts()
-        cur = self.layouts[win]
-        focus = cur.focus_next(win)
-        if not focus:
-            idx = layouts.index(cur)
-            while idx < len(layouts) - 1 and not focus:
-                idx += 1
-                focus = layouts[idx].focus_first()
-        return focus
-
-    def focus_previous(self, win):
-        layouts = self._get_layouts()
-        cur = self.layouts[win]
-        focus = cur.focus_previous(win)
-        if not focus:
-            idx = layouts.index(cur)
-            while idx > 0 and not focus:
-                idx -= 1
-                focus = layouts[idx].focus_last()
-        return focus
-
-    def __getattr__(self, name):
-        """Delegate unimplemented command calls to active layout.
-
-        For ``cmd_``-methods that don't exist on the Delegate subclass, this
-        looks for an implementation on the active layout.
-        """
-        if name.startswith('cmd_'):
-            return getattr(self._get_active_layout(), name)
-        return super().__getattr__(name)
-
-    @property
-    def commands(self):
-        return self._get_active_layout().commands
-
-    def info(self):
-        d = Layout.info(self)
-        for layout in self._get_layouts():
-            d[layout.name] = layout.info()
-        return d
-
-
 class _ClientList:
     """
     ClientList maintains a list of clients and a current client.
