+++ libqtile/layout/slice.py
@@ -27,7 +27,7 @@
 Slice layout. Serves as example of delegating layouts (or sublayouts)
 """
 
-from libqtile.layout.base import Delegate, Layout
+from libqtile.layout.base import Layout
 from libqtile.layout.max import Max
 
 
@@ -96,7 +96,7 @@ class Single(Layout):
         pass
 
 
-class Slice(Delegate):
+class Slice(Layout):
     """Slice layout
 
     This layout cuts piece of screen_rect and places a single window on that
@@ -112,7 +112,8 @@ class Slice(Delegate):
     ]
 
     def __init__(self, **config):
-        Delegate.__init__(self, **config)
+        self.layouts = {}
+        Layout.__init__(self, **config)
         self.add_defaults(Slice.defaults)
         self._slice = Single()
 
@@ -121,6 +122,26 @@ class Slice(Delegate):
         res._slice = self._slice.clone(group)
         res.fallback = self.fallback.clone(group)
         return res
+
+    def delegate_layout(self, windows, mapping):
+        """Delegates layouting actual windows
+
+        Parameters
+        ===========
+        windows:
+            windows to layout
+        mapping:
+            mapping from layout to ScreenRect for each layout
+        """
+        grouped = {}
+        for w in windows:
+            lay = self.layouts[w]
+            if lay in grouped:
+                grouped[lay].append(w)
+            else:
+                grouped[lay] = [w]
+        for lay, wins in grouped.items():
+            lay.layout(wins, mapping[lay])
 
     def layout(self, windows, screen_rect):
         win, sub = self._get_screen_rects(screen_rect)
@@ -167,8 +188,86 @@ class Slice(Delegate):
             self.fallback.add(win)
             self.layouts[win] = self.fallback
 
+    def remove(self, win):
+        lay = self.layouts.pop(win)
+        focus = lay.remove(win)
+        if not focus:
+            layouts = self._get_layouts()
+            idx = layouts.index(lay)
+            while idx < len(layouts) - 1 and not focus:
+                idx += 1
+                focus = layouts[idx].focus_first()
+        return focus
+
+    def hide(self):
+        for lay in self._get_layouts():
+            lay.hide()
+
+    def focus(self, win):
+        self.layouts[win].focus(win)
+
+    def blur(self):
+        for lay in self._get_layouts():
+            lay.blur()
+
+    def focus_first(self):
+        layouts = self._get_layouts()
+        for lay in layouts:
+            win = lay.focus_first()
+            if win:
+                return win
+
+    def focus_last(self):
+        layouts = self._get_layouts()
+        for lay in reversed(layouts):
+            win = lay.focus_last()
+            if win:
+                return win
+
+    def focus_next(self, win):
+        layouts = self._get_layouts()
+        cur = self.layouts[win]
+        focus = cur.focus_next(win)
+        if not focus:
+            idx = layouts.index(cur)
+            while idx < len(layouts) - 1 and not focus:
+                idx += 1
+                focus = layouts[idx].focus_first()
+        return focus
+
+    def focus_previous(self, win):
+        layouts = self._get_layouts()
+        cur = self.layouts[win]
+        focus = cur.focus_previous(win)
+        if not focus:
+            idx = layouts.index(cur)
+            while idx > 0 and not focus:
+                idx -= 1
+                focus = layouts[idx].focus_last()
+        return focus
+
+    def __getattr__(self, name):
+        """Delegate unimplemented command calls to active layout.
+
+        For ``cmd_``-methods that don't exist on the Slice class, this looks
+        for an implementation on the active layout.
+        """
+        if name.startswith('cmd_'):
+            return getattr(self._get_active_layout(), name)
+        return super().__getattr__(name)
+
     def cmd_next(self):
         self.fallback.cmd_next()
 
     def cmd_previous(self):
         self.fallback.cmd_previous()
+
+    @property
+    def commands(self):
+        return self._get_active_layout().commands
+
+    def info(self):
+        d = Layout.info(self)
+        for layout in self._get_layouts():
+            d[layout.name] = layout.info()
+        return d
